# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sumo Guard Health Analysis API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export type SleepLogQuality =
  (typeof SleepLogQuality)[keyof typeof SleepLogQuality];

export const SleepLogQuality = {
  poor: "poor",
  fair: "fair",
  good: "good",
  excellent: "excellent",
} as const;

export interface SleepLog {
  id: number;
  /** ISO datetime string of when user went to bed */
  bedtime: string;
  /** ISO datetime string of when user woke up */
  wakeTime: string;
  /** Calculated sleep duration in hours */
  durationHours: number;
  quality: SleepLogQuality;
  notes?: string | null;
  /** Date of the sleep entry (YYYY-MM-DD) */
  date: string;
  createdAt: string;
}

export type CreateSleepLogQuality =
  (typeof CreateSleepLogQuality)[keyof typeof CreateSleepLogQuality];

export const CreateSleepLogQuality = {
  poor: "poor",
  fair: "fair",
  good: "good",
  excellent: "excellent",
} as const;

export interface CreateSleepLog {
  /** ISO datetime string of when user went to bed */
  bedtime: string;
  /** ISO datetime string of when user woke up */
  wakeTime: string;
  quality: CreateSleepLogQuality;
  notes?: string | null;
  /** Date of the sleep entry (YYYY-MM-DD) */
  date: string;
}

export interface HabitLog {
  id: number;
  /** Date of the habit log (YYYY-MM-DD) */
  date: string;
  /** Minutes of exercise */
  exerciseMinutes: number;
  /** Glasses of water consumed */
  waterGlasses: number;
  /** Servings of fruits and vegetables */
  fruitVeggieServings: number;
  /** Hours of screen time */
  screenTimeHours: number;
  /** Stress level 1-10 */
  stressLevel: number;
  /** Number of cigarettes smoked */
  smokingCigarettes: number;
  /** Number of alcoholic drinks */
  alcoholDrinks: number;
  /** Minutes of meditation/mindfulness */
  meditationMinutes: number;
  notes?: string | null;
  createdAt: string;
}

export interface CreateHabitLog {
  /** Date of the habit log (YYYY-MM-DD) */
  date: string;
  /** Minutes of exercise */
  exerciseMinutes: number;
  /** Glasses of water consumed */
  waterGlasses: number;
  /** Servings of fruits and vegetables */
  fruitVeggieServings: number;
  /** Hours of screen time */
  screenTimeHours: number;
  /** Stress level 1-10 */
  stressLevel: number;
  /** Number of cigarettes smoked (default 0) */
  smokingCigarettes: number;
  /** Number of alcoholic drinks (default 0) */
  alcoholDrinks: number;
  /** Minutes of meditation/mindfulness */
  meditationMinutes: number;
  notes?: string | null;
}

export type DiseaseRiskRiskLevel =
  (typeof DiseaseRiskRiskLevel)[keyof typeof DiseaseRiskRiskLevel];

export const DiseaseRiskRiskLevel = {
  low: "low",
  moderate: "moderate",
  high: "high",
  critical: "critical",
} as const;

export interface DiseaseRisk {
  diseaseName: string;
  riskLevel: DiseaseRiskRiskLevel;
  /** Risk score 0-100 */
  riskScore: number;
  /** Predicted timeframe (e.g. "2-3 years", "5+ years") */
  predictedTimeframe: string;
  contributingFactors: string[];
  recommendations: string[];
}

export interface PredictionsResult {
  /** Overall health score 0-100 */
  overallHealthScore: number;
  riskAssessmentDate: string;
  diseases: DiseaseRisk[];
  summary: string;
  dataInsufficient: boolean;
}

export interface SleepTrendPoint {
  date: string;
  hours: number;
  quality: string;
}

export interface DashboardSummary {
  healthScore: number;
  avgSleepHours: number;
  avgExerciseMinutes: number;
  avgStressLevel: number;
  avgWaterGlasses: number;
  totalSleepLogs: number;
  totalHabitLogs: number;
  topRisks: DiseaseRisk[];
  sleepTrend: SleepTrendPoint[];
  recentLogs: HabitLog[];
}

export type UserProfileGender =
  (typeof UserProfileGender)[keyof typeof UserProfileGender];

export const UserProfileGender = {
  male: "male",
  female: "female",
  other: "other",
} as const;

export interface UserProfile {
  id: number;
  name: string;
  age: number;
  gender: UserProfileGender;
  /** Weight in kg */
  weight: number;
  /** Height in cm */
  height: number;
  existingConditions: string[];
  familyHistory: string[];
  createdAt: string;
}

export type UpdateUserProfileGender =
  (typeof UpdateUserProfileGender)[keyof typeof UpdateUserProfileGender];

export const UpdateUserProfileGender = {
  male: "male",
  female: "female",
  other: "other",
} as const;

export interface UpdateUserProfile {
  name: string;
  age: number;
  gender: UpdateUserProfileGender;
  weight: number;
  height: number;
  existingConditions: string[];
  familyHistory: string[];
}