# lib/api-zod/src/generated/types/createSleepLog.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sumo Guard Health Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateSleepLogQuality } from "./createSleepLogQuality";

export interface CreateSleepLog {
  /** ISO datetime string of when user went to bed */
  bedtime: string;
  /** ISO datetime string of when user woke up */
  wakeTime: string;
  quality: CreateSleepLogQuality;
  notes?: string | null;
  /** Date of the sleep entry (YYYY-MM-DD) */
  date: string;
}