# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sumo Guard Health Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { DiseaseRisk } from "./diseaseRisk";
import type { HabitLog } from "./habitLog";
import type { SleepTrendPoint } from "./sleepTrendPoint";

export interface DashboardSummary {
  healthScore: number;
  avgSleepHours: number;
  avgExerciseMinutes: number;
  avgStressLevel: number;
  avgWaterGlasses: number;
  totalSleepLogs: number;
  totalHabitLogs: number;
  topRisks: DiseaseRisk[];
  sleepTrend: SleepTrendPoint[];
  recentLogs: HabitLog[];
}