# lib/api-zod/src/generated/types/habitLog.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sumo Guard Health Analysis API
 * OpenAPI spec version: 0.1.0
 */

export interface HabitLog {
  id: number;
  /** Date of the habit log (YYYY-MM-DD) */
  date: string;
  /** Minutes of exercise */
  exerciseMinutes: number;
  /** Glasses of water consumed */
  waterGlasses: number;
  /** Servings of fruits and vegetables */
  fruitVeggieServings: number;
  /** Hours of screen time */
  screenTimeHours: number;
  /** Stress level 1-10 */
  stressLevel: number;
  /** Number of cigarettes smoked */
  smokingCigarettes: number;
  /** Number of alcoholic drinks */
  alcoholDrinks: number;
  /** Minutes of meditation/mindfulness */
  meditationMinutes: number;
  notes?: string | null;
  createdAt: string;
}