# lib/api-zod/src/generated/types/predictionsResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sumo Guard Health Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { DiseaseRisk } from "./diseaseRisk";

export interface PredictionsResult {
  /** Overall health score 0-100 */
  overallHealthScore: number;
  riskAssessmentDate: string;
  diseases: DiseaseRisk[];
  summary: string;
  dataInsufficient: boolean;
}