# lib/api-zod/src/generated/types/sleepLog.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sumo Guard Health Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { SleepLogQuality } from "./sleepLogQuality";

export interface SleepLog {
  id: number;
  /** ISO datetime string of when user went to bed */
  bedtime: string;
  /** ISO datetime string of when user woke up */
  wakeTime: string;
  /** Calculated sleep duration in hours */
  durationHours: number;
  quality: SleepLogQuality;
  notes?: string | null;
  /** Date of the sleep entry (YYYY-MM-DD) */
  date: string;
  createdAt: string;
}