# lib/api-zod/src/generated/types/sleepTrendPoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sumo Guard Health Analysis API
 * OpenAPI spec version: 0.1.0
 */

export interface SleepTrendPoint {
  date: string;
  hours: number;
  quality: string;
}