# lib/api-zod/src/generated/types/updateUserProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sumo Guard Health Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateUserProfileGender } from "./updateUserProfileGender";

export interface UpdateUserProfile {
  name: string;
  age: number;
  gender: UpdateUserProfileGender;
  weight: number;
  height: number;
  existingConditions: string[];
  familyHistory: string[];
}