# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sumo Guard Health Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { UserProfileGender } from "./userProfileGender";

export interface UserProfile {
  id: number;
  name: string;
  age: number;
  gender: UserProfileGender;
  /** Weight in kg */
  weight: number;
  /** Height in cm */
  height: number;
  existingConditions: string[];
  familyHistory: string[];
  createdAt: string;
}